Renaming ChannelTopic and GroupTopic types according to convention
