+++ slack-app-backend/src/main/java/com/slack/api/app_backend/events/payload/MessageChannelTopicPayload.java
@@ -1,0 +1,25 @@
+package com.slack.api.app_backend.events.payload;
+
+import com.slack.api.model.event.MessageChannelTopicEvent;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class MessageChannelTopicPayload implements EventsApiPayload<MessageChannelTopicEvent> {
+
+	private String token;
+	private String enterpriseId;
+	private String teamId;
+	private String apiAppId;
+	private String type;
+	private List<String> authedUsers;
+	private List<String> authedTeams;
+	private List<Authorization> authorizations;
+	private boolean isExtSharedChannel;
+	private String eventId;
+	private Integer eventTime;
+	private String eventContext;
+
+	private MessageChannelTopicEvent event;
+}
