+++ slack-app-backend/src/test/java/test_locally/sample_json_generation/EventsApiPayloadDumpTest.java
@@ -102,8 +102,8 @@ public class EventsApiPayloadDumpTest {
                 buildUserResourceDeniedPayload(),
                 buildUserResourceGrantedPayload(),
                 new UserResourceRemovedPayload(),
-                new GroupTopicMessagePayload(),
-                new ChannelTopicMessagePayload()
+                new MessageGroupTopicPayload(),
+                new MessageChannelTopicPayload()
         );
         for (EventsApiPayload<?> payload : payloads) {
             try {
